+++ app/Packaging.php
@@ -15,12 +15,21 @@ class Packaging extends Model
     protected $fillable = ['type','quantity','capacity'];
 
     public function getNameAttribute(){
-        return $this->quantity . ' ' . $this->type . ' x ' . $this->capacity/100 . ' l';
+        return $this->quantity . ' ' . $this->type . ' x ' . $this->capacity . ' l';
     }
-
 
     public function beers()
     {
         return $this->hasMany(Beer::class);
+    }
+
+    public function getCapacityAttribute($value)
+    {
+        return $value / 100;
+    }
+
+    public function setCapacityAttribute($value)
+    {
+        $this->attributes['capacity'] = $value * 100;
     }
 }